modified:   Repository/pom.xml modified:   Repository/src/main/java/virtualShop/dao/GenericDAO.java
new file:   Repository/src/main/java/virtualShop/dao/OrderDAO.java
new file:   Repository/src/main/java/virtualShop/dao/OrderProductDAO.java

new file:   Repository/src/main/java/virtualShop/dao/ProductDAO.java
new file:   Repository/src/main/java/virtualShop/dao/UserDAO.java
modified:   Repository/src/main/java/virtualShop/dao/impl/GenericDAOImpl.java
new file:   Repository/src/main/java/virtualShop/dao/impl/OrderDAOImpl.java
new file:   Repository/src/main/java/virtualShop/dao/impl/OrderProductDAOImpl.java

new file:   Repository/src/main/java/virtualShop/dao/impl/ProductDAOImpl.java

modified:   Repository/src/main/java/virtualShop/dao/impl/UserDAOImpl.java
new file:   Repository/src/main/java/virtualShop/entity/Application.java
modified:   Repository/src/main/java/virtualShop/entity/Order.java
new file:   Repository/src/main/java/virtualShop/entity/OrderProduct.java
modified:   Repository/src/main/java/virtualShop/entity/Product.java

modified:   Repository/src/main/java/virtualShop/entity/User.java
new file:   Repository/src/main/resources/META-INF/persistence.xml
modified:   Repository/src/main/resources/appContext-repository.xml
modified:   Service/pom.xml

new file:   Service/src/main/java/virtualShop/dto/OrderDTO.java
new file:   Service/src/main/java/virtualShop/dto/OrderProductDTO.java
new file:   Service/src/main/java/virtualShop/dto/ProductDTO.java
new file:   Service/src/main/java/virtualShop/dto/UserDTO.java
new file:   Service/src/main/java/virtualShop/dto/UserRole.java
new file:   Service/src/main/java/virtualShop/mapper/OrderMapper.java

new file:   Service/src/main/java/virtualShop/mapper/OrderProductMapper.java
new file:   Service/src/main/java/virtualShop/mapper/ProductMapper.java
new file:   Service/src/main/java/virtualShop/mapper/UserMapper.java
new file:   Service/src/main/java/virtualShop/service/OrderService.java
new file:   Service/src/main/java/virtualShop/service/UserService.java
new file:   Service/src/main/java/virtualShop/service/impl/OrderServiceImpl.java
new file:   Service/src/main/java/virtualShop/service/impl/UserServiceImpl.java
new file:   Service/src/main/resources/appContext-service.xml
modified:   Service/src/test/java/virtualShop/test/AppTest.java
modified:   Web/pom.xml
new file:   Web/src/main/webapp/WEB-INF/config/appContext-security.xml

new file:   Web/src/main/webapp/WEB-INF/config/appContext-web.xm
new file:   Web/src/main/webapp/WEB-INF/template/layout.jsp
new file:   Web/src/main/webapp/WEB-INF/tiles.xml
modified:   Web/src/main/webapp/WEB-INF/web.xml

+++ Repository/src/main/java/virtualShop/dao/GenericDAO.java
@@ -17,7 +17,7 @@ public interface GenericDAO<T extends DomainEntity> {
      * @param id criteria to search by
      * @return the matching object
      */
-    public T getObjectById(long id);
+    public T getObjectById(int id);
 
     /**
      * method to retrieve all objects from database
@@ -35,6 +35,6 @@ public interface GenericDAO<T extends DomainEntity> {
      * method to delete an object from database
      * @param id criteria to delete
      */
-    public void delete(long id);
+    public void delete(int id);
 
 }
